The least amout of syntax errors i've seen in ages
-cpu_map pins and constants are now short (can't cast int)
-Transferred more Parser methods into Controller
-removing parser soon

+++ Armanipulator_Firmware/cpu_map.h
@@ -11,27 +11,27 @@
 class Pinout {
 public:
 //Define Wrist Driver Pins
-	const static int WRIST_ROT_STEP = 2;
-	const static int WRIST_ROT_DIR = 3;
+	const static short WRIST_ROT_STEP = 2;
+	const static short WRIST_ROT_DIR = 3;
 //Define Grip Driver Pins
-	const static int GRIP_STEP = 4;
-	const static int GRIP_DIR = 5;
+	const static short GRIP_STEP = 4;
+	const static short GRIP_DIR = 5;
 
 //Define Arm Motion Pins
-	const static int EXTEND_STEP = 6;
-	const static int EXTEND_DIR = 7;
+	const static short EXTEND_STEP = 6;
+	const static short EXTEND_DIR = 7;
 
 //Define Microstep Control Pins
-	const static int MICROSTEP_1 = 8;
-	const static int MICROSTEP_2 = 9;
-	const static int MICROSTEP_3 = 10;
+	const static short MICROSTEP_1 = 8;
+	const static short MICROSTEP_2 = 9;
+	const static short MICROSTEP_3 = 10;
 };
 
 class MotorConfig {
 public:
-	const static int MOTOR_STEPS = 200;
-	const static int MICROSTEPS = 1;
-	const static int MOTOR_RPM = 120;
+	const static short MOTOR_STEPS = 200;
+	const static short MICROSTEPS = 1;
+	const static short MOTOR_RPM = 120;
 };
 
 #endif
